+++ src/components/LoginModal/LoginModal.tsx
@@ -3,7 +3,7 @@ import type { User } from 'firebase/auth'
 import { useToast } from '@entur/alert'
 import { Modal } from '@entur/modal'
 import { useUser } from '../../UserProvider'
-import { usePrevious } from '../../utils'
+import { usePrevious } from '../../hooks/usePrevious'
 import { EmailLogin } from './EmailLogin/EmailLogin'
 import { LoginOptions } from './LoginOptions/LoginOptions'
 import { Signup } from './Signup/Signup'
