+++ src/dashboards/Timeline/TimelineDashboard.tsx
@@ -11,7 +11,6 @@ import {
     getIconColor,
     timeUntil,
     getIconColorType,
-    usePrevious,
     isEqualUnsorted,
     BREAKPOINTS,
 } from '../../utils'
@@ -30,6 +29,7 @@ import {
 } from '../../settings/LocalStorage'
 import { LongPressProvider } from '../../logic/longPressContext'
 import { useCounter } from '../../hooks/useCounter'
+import { usePrevious } from '../../hooks/usePrevious'
 import './TimelineDashboard.scss'
 
 const TICKS = [-1, 0, 1, 2, 3, 4, 5, 10, 15, 20, 30, 60]
